consolidaitng different views into people index
