--- web/src/components/people/IAOView.module.css
@@ -1,8 +1,0 @@
-.button-container {
-  justify-self: end;
-  justify-items: end;
-  height: 4rem;
-  display: grid;
-  margin-bottom: 0.5rem;
-  margin-top: 1rem;
-}
