+++ web/src/pages/people/PeoplePage.module.css
@@ -8,3 +8,7 @@
   margin-bottom: 0.5rem;
   margin-top: 1rem;
 }
+
+.no-users-found {
+  text-align: center;
+}
