+++ web/src/pages/people/index.tsx
@@ -1,13 +1,42 @@
 import MetaHead from "@/components/meta/Head";
-import AdminView from "@/components/people/AdminView";
-import IAOView from "@/components/people/IAOView";
-import TreOpsStaffView from "@/components/people/TreOpsStaffView";
+import ApprovedResearcherImport from "@/components/people/ApprovedResearcherImport";
+import ExternalInvite from "@/components/people/ExternalInvite";
 import LoginFallback from "@/components/ui/LoginFallback";
 import Title from "@/components/ui/Title";
 import { useAuth } from "@/hooks/useAuth";
+import styles from "./PeoplePage.module.css";
+import { useEffect, useState } from "react";
+import { getUsers, UserData } from "@/openapi";
+import Box from "@/components/ui/Box";
+import { Alert, AlertMessage } from "@/components/shared/exports";
+import UserDataTable from "@/components/people/UserDataTable";
+import Callout from "@/components/ui/Callout";
 
 export default function PeoplePage() {
   const { authInProgress, isAuthed, userData } = useAuth();
+  const [users, setUsers] = useState<Array<UserData> | null>(null);
+  const [isLoading, setIsLoading] = useState(false);
+  const [errorMessage, setErrorMessage] = useState("");
+
+  useEffect(() => {
+    const fetchPeople = async () => {
+      setIsLoading(true);
+      try {
+        const response = await getUsers();
+        if (response.response.ok && response.data) {
+          setUsers(response.data);
+        }
+      } catch (error) {
+        console.error("Failed to get people:", error);
+        setErrorMessage("Failed to get people");
+        setUsers(null);
+      } finally {
+        setIsLoading(false);
+      }
+    };
+    fetchPeople();
+  }, []);
+
   if (authInProgress) return null;
 
   if (!isAuthed) return <LoginFallback />;
@@ -15,6 +44,25 @@ export default function PeoplePage() {
   const isAdmin = userData?.roles.includes("admin");
   const isIAO = userData?.roles.includes("information-asset-owner");
   const isTreOpsStaff = userData?.roles.includes("tre-ops-staff");
+
+  if (!isAdmin && !isTreOpsStaff && !isIAO)
+    return (
+      <Alert type="warning">
+        <AlertMessage>You do not have permission to view this page</AlertMessage>
+      </Alert>
+    );
+
+  if (!users || users.length === 0)
+    return (
+      <Box>
+        <div className={styles["no-users-found"]}>No users found</div>
+        {errorMessage && (
+          <Alert type="error">
+            <AlertMessage>{errorMessage}</AlertMessage>
+          </Alert>
+        )}
+      </Box>
+    );
 
   return (
     <>
@@ -35,7 +83,15 @@ export default function PeoplePage() {
                 : "You do not have permission to view this page"
         }
       />
-      {isAdmin ? <AdminView /> : isTreOpsStaff ? <TreOpsStaffView /> : isIAO && <IAOView />}
+
+      {!isTreOpsStaff && (
+        <div className={styles["button-container"]}>
+          {isAdmin && <ApprovedResearcherImport />}
+          {(isAdmin || isIAO) && <ExternalInvite />}
+        </div>
+      )}
+      {!isAdmin && <Callout construction />}
+      <UserDataTable canEdit={true} users={users} setUsers={setUsers} isLoading={isLoading} />
     </>
   );
 }
